pylint: fix baseldap/privilege cyclic import
enforce_managed_permissions() imported principal_has_privilege from
privilege.py, which already imports baseldap at module load time.
Pylint reports R0401 (cyclic-import) and make lint exits with error 8
since fa2555cfd landed on master.

Move principal_has_privilege() to privilege_util.py. baseldap and
privilege import it from there without creating a cycle.

Also stub the pki namespace in vault.py for the non-server branch,
matching a65ebebd3, and drop the now-useless used-before-assignment
suppression in vault_del.post_callback.
